refactor(commands): extract /chrome command into chrome.rs
